refactor(app): #373 Phase 1-9 commands を lib/app-commands.ts へ切り出し
App.tsx 1431 -> 1275 行 (-156)。コマンドパレット用 Command[] 構築
(useMemo 本体 + recentProjects/THEMES_FOR_PALETTE の動的展開) を新規
ファイル src/renderer/src/lib/app-commands.ts の pure 関数
buildAppCommands に集約。

設計:
- pure 関数 (lib/team-prompts.ts と同じ系譜)。副作用なしで Command[] を
  return するだけ。run / when クロージャは引数を closure-capture する
- ファイル位置: src/renderer/src/lib/app-commands.ts (引き継ぎ書通り、
  hooks/ 配下ではない)
- useMemo の deps 配列は **App.tsx 呼び出し側に残す**: react-hooks/
  exhaustive-deps が呼び出し側で機能する形を維持。pure 関数なので hook
  をかぶせる必要はない (Phase 1-1 〜 1-7 の useEffect 持ち hook とは性質
  が違う)
- THEMES_FOR_PALETTE 定数を app-commands.ts に移動 (App.tsx で他に参照
  なし、grep で確認済み)

deps 設計:
- 配列 (recentlyClosed / terminalTabs / diffTabs) を生で渡さず、
  recentlyClosedLength / terminalTabsLength / diffTabsLength と
  settings.recentProjects (slice 用) のみ受ける形にして useMemo の
  identity 振動を防ぐ
- settings 全体ではなく、コマンドパレットが実際に使う 4 項目
  (theme / density / recentProjects / language) のみ受ける slice 型
- maxTerminals は use-terminal-tabs.ts の MAX_TERMINALS を呼び出し側
  から渡す (hook 戻り値ではないため)

不変式 (Issue #373 全 Phase 共通) を保持:
- IPC コマンド名・event 名は一切変更なし
- コマンド ID (project.new / theme.{name} / app.checkForUpdates 等) は
  逐字保持
- t() による i18n キー解決タイミングは現状と同一 (build 時)
- when() / run() のクロージャ依存も現状と同一
- dynamic import パスは ./lib/updater-check → ./updater-check に
  hook ファイル位置に合わせて訂正 (機能は同一)

スコープ判断:
- 引き継ぎ書記載の「JSX サブツリーを子 component 化」は **Phase 1-9 で
  は同梱しない**: 1 PR = 1 関心事 を維持し、PR 肥大化を回避。Phase 1-10
  (Terminal リファクタ) または Phase 4 (CanvasLayout / SettingsModal
  細分化) と一緒に扱う方針

検証:
- npm run typecheck: green
- cargo check --manifest-path src-tauri/Cargo.toml: green
- cargo clippy --no-deps: 既存ベースライン 15 件のまま (新規警告ゼロ)

累計削減 (Phase 1-1 〜 1-9): -1520 行 (#373 全体目標 -1995, 進捗 76%)。

Refs #373
